project/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
PyJWT==2.8.0
PyJWT==2.8.0
python_bcrypt==0.3.2
